Update Contacts when GUI is changed
Current implementation during editing of AB3 Persons is that
all with the previous name will be marked as not in AB3, while tasks
with the new edited name will be marked as in AB3. The names in Contact
will NOT be updated to the edited name as of this commit.